requirements.txt
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==38.0.4
Flask==2.2.2
Flask-Cors==3.0.10
Flask-SQLAlchemy==2.5.1
gevent==22.10.2
greenlet==2.0.1
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
Naked==0.1.32
pycparser==2.21
pycryptodome==3.15.0
PyMySQL==1.0.2
PyYAML==6.0
requests==2.28.1
shellescape==3.8.1
six==1.16.0
SQLAlchemy==1.4.44
urllib3==1.26.12
Werkzeug==2.2.2
zope.event==4.5.0
zope.interface==5.5.2
